For NTPsec, ignore error about statistics directory not existing (#10502)
NTPsec by default wants to log statistics to a file (or at least check
to see that the statistics directory exists, even if no logging is
enabled). To match the behavior we had with NTP where no logging is
enabled, don't configure the statistics directory for NTPsec, and ignore
any errors about it not existing.

Signed-off-by: Saikrishna Arcot <sarcot@microsoft.com>
Change-Id: Ie5c791585a2467c7b622c95cb1be03a1745ba56b