Update the watchdog timeout for the Mellanox platform.
Remove the redundancy definition,
Increase the valid watchdog timeout, due to some network issue, 10s timeout for the test is not enough, increase the timeout to make the test muck more stable.
For bulldog and boxer platform, the max watchdog is 32, so could not use the same value as other platforms.

Change-Id: I09d894beda5cfca7532c047c5f26cb82a618443f