Relax the check on the number of PORT_READY messages (#11399)
Sometimes, after fast reboot, there is a random number of PORT_READY
messages that occur; it could be 0, it could be the number of ports that
are up, it could be some number in between. The reasoning for this is a
bit unclear to me. However, considering the logic here prior to my
changes had a relaxed check here, restore that behavior.

Signed-off-by: Saikrishna Arcot <sarcot@microsoft.com>
Change-Id: Ie77891350253de7b045c3f91e3e4e2a6d2d16122
(cherry picked from commit 9fe283009a00038bc565db207c1607ff1a6e0eb4)