Fix Auto FEC Validation
Details:
A successful auto fec configuration will result in the link being up with any fec mode,
Negotiating FEC is not always deterministic, so we have no way to determine in the validation the expected fec.
when configuring auto fec, any fec mode the ports have negotiated is valid.

This is fix for bug #3788218.

Change-Id: Ifcde6287120aa8bdcdf1404dae3b50395701d052