Update test to support new GNMI path schema (#11780)
What is the motivation for this PR?
We have updated database configuration to support multiple databases for smartswitch.
And then we need to update sonic-gnmi and sonic-mgmt test to support new database configuration and new GNMI path for smartswitch.

How did you do it?
Update test for gnmi and dash, add localhost in GNMI path.

How did you verify/test it?
Run end to end test.

Change-Id: Idc12a940f19b3dd9097a2502b8344f391dba8f00