Update the pfcwd test cases according to the new implement of the pfcwd.
After the feature is completed, it criteria to trigger the pfcwd is much more strict than previous on Nvidia platform, So adjust the test case to properly trigger the pfcwd. The solution is to send background traffic during the pfc frame is sent from the fanout.

Feature: Feature #3636844: Customers | MSFT | Support long-term changes to trigger PFC WD

Change-Id: Icdc89e9d30a505433dde76c95da5951d80931c6c