Filtered non-supported cables from auto-neg tests
Details:
Skipped auto-neg tests on known unsupported cable types (see class of Cable_Compliance_Const in ngts/constants/constants.py).
Whenever a new cable type is added to a setup, we should adjust the test to work on this type, whether the new cable should be skipped on the auto-neg tests or not.
In order to determine that, we can look at the Vendor PN value of the port (as can be seen in show interfaces transceiver eeprom port_name) and look online if it should be skipped on auto-neg (for example, if it's an optic cable).

Change-Id: I11b2fa94f82bfaeb6c571b2ac447b55146cbe008