Bug Handler: need to use use flat=True when fetch the dump file from DUT
Here is the explanation of flat in the fetch module:
Allows you to override the default behavior of appending hostname/path/to/file to the destination.
If dest ends with ‘/’, it will use the basename of the source file, similar to the copy module.
This can be useful if working with a single host, or if retrieving files that are uniquely named per host.
If using multiple hosts with the same filename, the file will be overwritten for each host.
Choices:
false ← (default)
true

Change-Id: Idd24c570d929ea94fe52fd4cb85e64650c9c9f7a