Fix random seed issue
It will fix the issue: When a test set has a function-level fixture which generates a random value for each sub-test case, it will always get the same random value for the test set.
Solution: The random seed will be constructed with the random seed of the session level and the hash value of each sub-test name. So it can make sure the different test has different random seed, and with the same random seed of the session level whatever run one single case, or runs all cases together, we get the same result.

Change-Id: I3c4eb77944ca2156e2e61ba9f087d8c6c190ea80