Adjust the vxlan ecmp cases file
The include_long_tests tests can not run together with other tests due to a test issue in the setUp fixture scope.
Why not fix the issue? Because there is another pytest issue, after change the scope of setUp to "class", it will be executed for every test case, which will increase the test run time by about 1 hour for the full test. The pytest issue is caused by parameterization. So the simplest way to fix the test failure is running the long tests separately.

Change-Id: I56163cb1150a037791864e09b146d03b3ac3fd08