Update the bug handler config file
Update the bug handler config file to let bug handler have capability to reopen the bug when detected a closed bug.
After the change, it will reopen the ticket when it is in status: "Closed(5)", "Closed (Rejected)(38)" to Assigned, and by default it will change the assignee to Moshe(11866). To check the id of ticket status or the assignee, you can use the link: https://redmine.mellanox.com/issues/<ticket number>.xml

Change-Id: I66f872e53eaf6481ddca9bbba7927799c704c7e8