Fix device_vlan_map_list generation issue in library conn_graph_facts
When some port with a bunch of vlan ids in fact port_vlans
Current device_vlan_map_list generation algorithm would miss some port index
If the vlan of port A matched the port B which with a bunch of vlan ids, the
index of port B would be picked into the device_vlan_map_list.
Then the correct port A would failed to be picked in device_vlan_map_list.

The solution is make sure the port which with a bunch of vlan ids is always in the last position of port_name_sorted_by_value.
So that the port pick process would not be affected.

Change-Id: Iedc45814d9328836b6e2c057473cbfe35824b5ca
(cherry picked from commit 68d90eba115b3f7af40e496b5e5589f2f162b050)