fix dynamic buffer test issue
For nvidia device, the headroom size is related to the speed and cable length. In community setup, some cable length is 300m in config, so when the speed is over 400G, the headroom size will exceed the limitation.
So for test_shared_headroom_pool_configure, we change the cable length to a smaller when speed is equal or greater 400G, and the cable length is over 200m.

Change-Id: I1b23725f42bb379ff4519136c408cb0460e67123