do not load the dynamic err msg when loganalyzer is disabled
In the previous implement, loading the dynamic ignored err msg need to happens always, in some team, they don't have access to the Redmine, so it will throw exception when getting the dynamic ignored err msg.

So add condition for loading the dynamic ignored err msg: when the loganalyzer is enabled or user want to always load the err msg by specifying an option.

Change-Id: I8c2a12a691bc8d4d5cfcf713664048c94e554bef