Not run loganalyzer in subprocess when only one dut is in the test
When run the loganalyzer in parallel, the log in the subprocess could not be uploaded to the allure report. it is inconvenient for us to debug issues for the bug handler.
This is a temp solution, long term solution should be make all the log in the subprocess also uploaded to the allure report.

Change-Id: I005b6dc39d2af071a07fab7d8c0662fb99b3fd91