Removing warning msg check for TC status, because warning msg is expected according to HW's feedback as below
On SPC systems PWM is controlled by the ASIC (For SPC2,SPC3… controlled by CPLD). When you are restarting SDK – pwm is removing till the SDK start and create PWM sysfs attribute again. For Thermal control missing PWM it’s a serious error because it can’t control FAN’s. In this case TC reporting error when SDK is not started.
TC will be recovered when SDK start and can be controlled PWM.You should ignore this messages on SPC1/SIB2 systems on system boot or SDK stop.

Change-Id: If657535485c1ebf72426071126240f02e473a5c8